+++ fluid/Fluid.h
@@ -8,14 +8,14 @@
 
 
 // Fluid magic numbers
-const double FluidTimestep = 0.005f;
-const double FluidSmoothLen = 0.012f;
-const double FluidStaticStiff = 3000.0f;
-const double FluidRestDensity = 1000.0f;
-const double FluidWaterMass = 0.0002f;
-const double FluidViscosity = 0.1f;
-const double FluidStiff = 200.0f;
-const double FluidInitialSpacing = 0.0045f;
+const float FluidTimestep = 0.005f;
+const float FluidSmoothLen = 0.012f;
+const float FluidStaticStiff = 3000.0f;
+const float FluidRestDensity = 1000.0f;
+const float FluidWaterMass = 0.0002f;
+const float FluidViscosity = 0.1f;
+const float FluidStiff = 200.0f;
+const float FluidInitialSpacing = 0.0045f;
 
 /*****************************************************************************/
 
@@ -23,7 +23,7 @@ struct FluidNeighborRecord
 {
 	unsigned int p; // Particle Index
 	unsigned int n; // Neighbor Index		
-	double distsq; // Distance Squared and Non-Squared (After Density Calc)
+	float distsq; // Distance Squared and Non-Squared (After Density Calc)
 };
 
 struct FluidGridOffset 
@@ -39,8 +39,8 @@ struct Particle
 	D3DXVECTOR2 pos;
 	D3DXVECTOR2 vel;
 	D3DXVECTOR2 acc;
-	double density = 0;
-	double pressure = 0;
+	float density = 0;
+	float pressure = 0;
 };
 
 class Fluid 
@@ -50,10 +50,10 @@ class Fluid
 		Fluid();
 		~Fluid();
 
-		void Create(double w, double h);
-		void Fill(double size);
+		void Create(float w, float h);
+		void Fill(float size);
 		void Clear();
-		void Update(double dt);
+		void Update(float dt);
 
 		/* Common Data */
 		unsigned int * gridindices;
@@ -68,8 +68,8 @@ class Fluid
 		unsigned int Step()					{ return step; }
 		void Pause( bool p )				{ paused = p; }
 		void PauseOnStep( unsigned int p )	{ pause_step = p; }
-		double Width()						{ return width; }
-		double Height()						{ return height; }
+		float Width()						{ return width; }
+		float Height()						{ return height; }
 
 	private:
 		
@@ -80,7 +80,7 @@ class Fluid
 		void ComputeDensity();
 		void SqrtDist();
 		void ComputeForce();
-		void Integrate(double dt);
+		void Integrate(float dt);
 
 	private:
 		/* Run State */
@@ -89,13 +89,13 @@ class Fluid
 		unsigned int pause_step;
 
 		/* World Size */
-		double width;
-		double height;
+		float width;
+		float height;
 		int grid_w;
 		int grid_h;
 
 		/* Coefficients for kernel */
-		double poly6_coef;
-		double grad_spiky_coef;
-		double lap_vis_coef;
+		float poly6_coef;
+		float grad_spiky_coef;
+		float lap_vis_coef;
 };
